(front): rename bookContent.ts to bookContext.ts
