+++ frontend/app/(protected)/(drawer)/(tabs)/library.tsx
@@ -18,7 +18,7 @@ import { StackNavigationProp } from "@react-navigation/stack";
 import { RootStackParamList } from "./types"; // Import your defined types
 import Loading from "@/components/Loading";
 import { AuthContext, type User } from "@/utilities/authContext";
-import { BookContext } from "@/utilities/bookContent";
+import { BookContext } from "@/utilities/bookContext";
 import { getAllBooks, uploadBookToDB } from "@/utilities/backendService";
 
 const { width } = Dimensions.get("window");
